Add Magic Description
This commit adds the class Magic.Description, which holds information of
a spell, such as cost and name. This description is used in the
spellbook gui. With this addition, every magic now needs to know its
name, which is why every magic class was slightly changed to add a
constructor taking a String.