Copy capabilities on death
This commit fixes a but where capabilities (such as staff GUI animations toggle and
lodestone magic teleport location) were not copied on death (or end ->
overworld teleportation). This was due to the fact that player
capabilities are not copied by default on death, but has to be done
manually via the PlayerEvent.Clone event.