Add De-age magic
This commit adds a new small spell, the de-age magic. This spell
functions similar to the egg magic spell, but instead of creating spawn
eggs, it turns mobs into their child variants. Due to the similarity with
the egg magic, a common superclass "RayMagic" was created to avoid code
duplication.