Change MapMagic to find villages
This commit changes the MapMagic from finding monuments and mansions to
finding nearby villages. This change was done because the server would
quite often hang for ~10 seconds trying to find a mansion. If I come up
with a solution that can find mansions without hanging the server, I
might change the spell back, but this will have to do for now.