Rework BlueDyeMagic
Before, the BlueDyeMagic was implemented similar to how the weather
command works. This was however not ideal, since the weather command
works in a weird way (for example, with '/weather clear' it will always
start to both rain and thunder after 5 minutes). This should no longer
happen, since the new implementation of the spell works more in synch
with the weather cycle, and does not rely on the clearWeatherTime
variable.