Separate client and server Magics instances
This commit creates two separate instances of Magics, one for the
logical client and one for the logical server. This means that the
UpdateMagicsMessage no longer overwrites the server magics in
singleplayer, when both logical sides are on same JVM. See [1] for
more information, especially about the use of EffectiveSide.get()
in WizardStaffItem.getUseDuration().

[1] https://forums.minecraftforge.net/topic/98910-jsonreloadlistener-clientserver-separation/